fix(rust): preserve policy after upstream rebase
