+++ zeroshot-rust/src/native_v2_codex/tests.rs
@@ -1,5 +1,7 @@
 #![cfg(unix)]
 
+#[path = "tests/cancellation.rs"]
+mod cancellation;
 #[path = "tests/correction.rs"]
 mod correction;
 #[path = "tests/environment.rs"]
@@ -11,9 +13,8 @@ mod retry;
 
 use std::collections::{BTreeMap, BTreeSet};
 use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::sync::Arc;
-use std::time::Duration;
 
 use openengine_cluster_protocol::{IdempotencyKey, NodeName, RunSize, RunTitle, WorkerOutcome};
 use openengine_cluster_testkit::assertions::AssertValue;
@@ -271,7 +272,7 @@ fn assert_openrouter_capture(capture: &str) {
             "missing capture evidence: {expected}"
         );
     }
-    assert_schema_capture(&capture);
+    assert_schema_capture(capture);
     for suppressed in ["--ignore-user-config", "--ignore-rules", "--strict-config"] {
         assert!(!capture.contains(suppressed));
     }
@@ -420,61 +421,4 @@ async fn malformed_output_stops_after_two_correction_turns() {
     let capture = fs::read_to_string(capture).assert_value();
     assert_eq!(capture.matches("prompt=").count(), 3);
     assert_eq!(capture.matches("arg=resume").count(), 2);
-}
-
-#[tokio::test]
-async fn cancellation_waits_for_contained_child_cleanup() {
-    let directory = TestDirectory::new("codex-cancel");
-    let capture = directory.child("capture");
-    let pid_path = directory.child("pid");
-    let (admitted, runtime) = openai_runtime(
-        &directory,
-        SessionScope::Execution,
-        &["CAPTURE_PATH", "CODEX_API_KEY", "PID_PATH", "SLOW_RUN"],
-    )
-    .await;
-    let mut handle = start(
-        &runtime,
-        &admitted,
-        1,
-        &[
-            ("CAPTURE_PATH", capture.display().to_string()),
-            ("CODEX_API_KEY", "fake-openai-key".to_owned()),
-            ("PID_PATH", pid_path.display().to_string()),
-            ("SLOW_RUN", "true".to_owned()),
-        ],
-    )
-    .await;
-    let mut attach = handle.take_initial_output().assert_value();
-    assert_eq!(
-        attach.recv_output().await.assert_value().text,
-        "Codex turn started"
-    );
-    assert_eq!(
-        attach.recv_output().await.assert_value().text,
-        r#"{"response":{"answer":42}}"#
-    );
-    let pid = wait_for_pid(&pid_path).await;
-    handle.cancel();
-    assert_eq!(handle.completion().await, Err(NodeRunnerError::Cancelled));
-    assert!(
-        !process_is_live(pid),
-        "provider child remained alive after completion"
-    );
-}
-
-async fn wait_for_pid(path: &Path) -> u32 {
-    for _ in 0..100 {
-        if let Ok(value) = fs::read_to_string(path) {
-            if let Ok(pid) = value.trim().parse() {
-                return pid;
-            }
-        }
-        tokio::time::sleep(Duration::from_millis(10)).await;
-    }
-    None::<u32>.assert_value_with("script did not publish its pid")
-}
-
-fn process_is_live(pid: u32) -> bool {
-    PathBuf::from(format!("/proc/{pid}")).exists()
 }
